refactor - 사다리인터페이스 추가.
