+++ src/test/java/nextstep/OnlineLadderTest.java
@@ -5,11 +5,11 @@ import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-class LadderTest {
+class OnlineLadderTest {
     @Test
     @DisplayName("게임?")
     void game() {
-        Ladder ladder = new Ladder(2);
+        OnlineLadder ladder = new OnlineLadder(2);
         ladder.add("bong");
         ladder.add("big");
         ladder.add("cptbong");
